EDU-1115 Extract common form inputs for track payers
