+++ src/plugins/media-slideshow/media-slideshow-editor.js
@@ -15,13 +15,12 @@ import Timeline from '../../components/media-player/timeline.js';
 import { formatMediaPosition } from '../../utils/media-utils.js';
 import { useService } from '../../components/container-context.js';
 import { sectionEditorProps } from '../../ui/default-prop-types.js';
-import ObjectWidthSlider from '../../components/object-width-slider.js';
 import MediaPlayer from '../../components/media-player/media-player.js';
+import TrackEditor from '../../components/media-player/track-editor.js';
 import { usePercentageFormat } from '../../components/locale-context.js';
 import { ensureIsExcluded, removeItemAt } from '../../utils/array-utils.js';
-import MainTrackEditor from '../../components/media-player/main-track-editor.js';
 import { useMediaDurations } from '../../components/media-player/media-hooks.js';
-import MediaVolumeSlider from '../../components/media-player/media-volume-slider.js';
+import PlayerSettingsEditor from '../../components/media-player/player-settings-editor.js';
 import { FORM_ITEM_LAYOUT, MEDIA_SCREEN_MODE, SOURCE_TYPE } from '../../domain/constants.js';
 import { createCopyrightForSourceMetadata, getAccessibleUrl } from '../../utils/source-utils.js';
 
@@ -40,7 +39,7 @@ function MediaSlideshowEditor({ content, onContentChanged }) {
   const [selectedChapterIndex, setSelectedChapterIndex] = useState(0);
   const [selectedChapterFraction, setSelectedChapterFraction] = useState(0);
 
-  const { sourceUrl, playbackRange, width, initialVolume, chapters } = content;
+  const { sourceUrl, playbackRange, initialVolume, chapters } = content;
 
   const [mediaDuration] = useMediaDurations([getAccessibleUrl({ url: sourceUrl, cdnRootUrl: clientConfig.cdnRootUrl })]);
   const sourceDuration = mediaDuration.duration;
@@ -57,8 +56,12 @@ function MediaSlideshowEditor({ content, onContentChanged }) {
     onContentChanged(newContent);
   };
 
-  const handleMainTrackContentChange = changedContent => {
-    changeContent({ ...changedContent });
+  const handleTrackContentChange = changedContent => {
+    changeContent(changedContent);
+  };
+
+  const handlePlayerSettingsContentChange = changedContent => {
+    changeContent(changedContent);
   };
 
   const handleChapterStartPositionChange = (key, newStartPosition) => {
@@ -66,14 +69,6 @@ function MediaSlideshowEditor({ content, onContentChanged }) {
     chapter.startPosition = newStartPosition;
     const newChapters = [...chapters];
     changeContent({ chapters: newChapters });
-  };
-
-  const handleWidthChanged = newValue => {
-    changeContent({ width: newValue });
-  };
-
-  const handleInitialVolumeChange = newValue => {
-    changeContent({ initialVolume: newValue });
   };
 
   const handleChapterAdd = startPosition => {
@@ -178,27 +173,18 @@ function MediaSlideshowEditor({ content, onContentChanged }) {
   return (
     <div className="MediaSlideshowEditor">
       <Form layout="horizontal" labelAlign="left">
-        <MainTrackEditor
+        <TrackEditor
+          content={content}
+          useName={false}
+          onContentChange={handleTrackContentChange}
+          />
+        <PlayerSettingsEditor
           content={content}
           useShowVideo={false}
           useAspectRatio={false}
           usePosterImage={false}
-          onContentChanged={handleMainTrackContentChange}
+          onContentChange={handlePlayerSettingsContentChange}
           />
-        <FormItem
-          label={<Info tooltip={t('common:widthInfo')}>{t('common:width')}</Info>}
-          {...FORM_ITEM_LAYOUT}
-          >
-          <ObjectWidthSlider value={width} onChange={handleWidthChanged} />
-        </FormItem>
-        <FormItem label={t('common:initialVolume')} {...FORM_ITEM_LAYOUT} >
-          <MediaVolumeSlider
-            value={initialVolume}
-            useValueLabel
-            useButton={false}
-            onChange={handleInitialVolumeChange}
-            />
-        </FormItem>
 
         <div className="MediaSlideshowEditor-playerPreview">
           <div className="MediaSlideshowEditor-playerPreviewLabel">{t('common:preview')}</div>
