+++ src/plugins/multitrack-media/multitrack-media-editor.js
@@ -1,40 +1,29 @@
-import Info from '../../components/info.js';
+import { Button, Form } from 'antd';
 import { useTranslation } from 'react-i18next';
 import React, { useMemo, useState } from 'react';
 import { PlusOutlined } from '@ant-design/icons';
 import cloneDeep from '../../utils/clone-deep.js';
-import { Button, Form, Radio, Switch } from 'antd';
-import UrlInput from '../../components/url-input.js';
 import ItemPanel from '../../components/item-panel.js';
 import ClientConfig from '../../bootstrap/client-config.js';
-import { analyzeMediaUrl } from '../../utils/media-utils.js';
 import { getAccessibleUrl } from '../../utils/source-utils.js';
+import { shouldDisableVideo } from '../../utils/media-utils.js';
 import { createDefaultTrack } from './multitrack-media-utils.js';
 import { useService } from '../../components/container-context.js';
 import { sectionEditorProps } from '../../ui/default-prop-types.js';
+import { removeItemAt, swapItemsAt } from '../../utils/array-utils.js';
 import TrackEditor from '../../components/media-player/track-editor.js';
-import ObjectWidthSlider from '../../components/object-width-slider.js';
 import TrackMixerEditor from '../../components/media-player/track-mixer-editor.js';
-import MediaVolumeSlider from '../../components/media-player/media-volume-slider.js';
-import { ensureIsExcluded, removeItemAt, swapItemsAt } from '../../utils/array-utils.js';
+import PlayerSettingsEditor from '../../components/media-player/player-settings-editor.js';
 import MultitrackMediaPlayer from '../../components/media-player/multitrack-media-player.js';
-import { FORM_ITEM_LAYOUT, MEDIA_ASPECT_RATIO, RESOURCE_TYPE, SOURCE_TYPE } from '../../domain/constants.js';
-
-const FormItem = Form.Item;
-const RadioGroup = Radio.Group;
-const RadioButton = Radio.Button;
-
-const shouldDisableVideo = sourceUrl => {
-  const { resourceType } = analyzeMediaUrl(sourceUrl);
-  return ![RESOURCE_TYPE.video, RESOURCE_TYPE.none, RESOURCE_TYPE.unknown].includes(resourceType);
-};
 
 function MultitrackMediaEditor({ content, onContentChanged }) {
   const clientConfig = useService(ClientConfig);
   const { t } = useTranslation('multitrackMedia');
-  const [selectedVolumePresetIndex, setSelectedVolumePresetIndex] = useState(0);
 
-  const { tracks, volumePresets, showVideo, aspectRatio, posterImage, width, initialVolume } = content;
+  const [selectedVolumePresetIndex, setSelectedVolumePresetIndex] = useState(0);
+  const [disableVideo, setDisableVideo] = useState(shouldDisableVideo(content.tracks[0].sourceUrl));
+
+  const { tracks, volumePresets, showVideo, aspectRatio, posterImage, initialVolume } = content;
 
   const sources = useMemo(() => {
     return tracks.map(track => ({
@@ -45,47 +34,20 @@ function MultitrackMediaEditor({ content, onContentChanged }) {
 
   const changeContent = newContentValues => {
     const newContent = { ...content, ...newContentValues };
+    const newDisableVideo = shouldDisableVideo(newContent.tracks[0].sourceUrl);
+
+    const reEnableVideo = !content.showVideo;
+    newContent.showVideo = newDisableVideo ? false : reEnableVideo || newContent.showVideo;
+    newContent.posterImage = newDisableVideo ? { sourceUrl: '' } : newContent.posterImage;
+
+    setDisableVideo(newDisableVideo);
     onContentChanged(newContent);
   };
 
-  const handeTrackContentChanged = (index, value) => {
-    const isMainTrack = index === 0;
+  const handeTrackContentChange = (index, value) => {
     const newTracks = cloneDeep(tracks);
     newTracks[index] = value;
-
-    const disableVideo = isMainTrack && shouldDisableVideo(value.sourceUrl);
-    if (disableVideo) {
-      changeContent({ tracks: newTracks, showVideo: false, posterImage: { sourceUrl: '' } });
-      return;
-    }
-
     changeContent({ tracks: newTracks });
-  };
-
-  const handleAspectRatioChanged = event => {
-    changeContent({ aspectRatio: event.target.value });
-  };
-
-  const handleShowVideoChanged = value => {
-    const newContent = { showVideo: value };
-
-    if (!newContent.showVideo) {
-      newContent.posterImage = { sourceUrl: '' };
-    }
-
-    changeContent(newContent);
-  };
-
-  const handlePosterImageSourceUrlChange = url => {
-    changeContent({ posterImage: { sourceUrl: url } });
-  };
-
-  const handleWidthChanged = newValue => {
-    changeContent({ width: newValue });
-  };
-
-  const handleInitialVolumeChange = newValue => {
-    changeContent({ initialVolume: newValue });
   };
 
   const handleMoveTrackUp = index => {
@@ -123,6 +85,10 @@ function MultitrackMediaEditor({ content, onContentChanged }) {
     changeContent({ tracks: newTracks, volumePresets: newVolumePresets });
   };
 
+  const handlePlayerSettingsContentChange = changedContent => {
+    changeContent(changedContent);
+  };
+
   const handleSelectedVolumePresetChange = volumePresetIndex => {
     setSelectedVolumePresetIndex(volumePresetIndex);
   };
@@ -141,8 +107,7 @@ function MultitrackMediaEditor({ content, onContentChanged }) {
           >
           <TrackEditor
             content={track}
-            usePlaybackRange
-            onContentChanged={value => handeTrackContentChanged(trackIndex, value)}
+            onContentChange={value => handeTrackContentChange(trackIndex, value)}
             />
         </ItemPanel>
       );
@@ -168,7 +133,8 @@ function MultitrackMediaEditor({ content, onContentChanged }) {
         >
         <TrackEditor
           content={track}
-          onContentChanged={value => handeTrackContentChanged(trackIndex, value)}
+          usePlaybackRange={false}
+          onContentChange={value => handeTrackContentChange(trackIndex, value)}
           />
       </ItemPanel>
     );
@@ -184,51 +150,11 @@ function MultitrackMediaEditor({ content, onContentChanged }) {
         </Button>
 
         <ItemPanel header={t('common:player')}>
-          <FormItem label={t('common:aspectRatio')} {...FORM_ITEM_LAYOUT}>
-            <RadioGroup
-              value={aspectRatio}
-              defaultValue={MEDIA_ASPECT_RATIO.sixteenToNine}
-              disabled={shouldDisableVideo(tracks[0].sourceUrl)}
-              onChange={handleAspectRatioChanged}
-              >
-              {Object.values(MEDIA_ASPECT_RATIO).map(ratio => (
-                <RadioButton key={ratio} value={ratio}>{ratio}</RadioButton>
-              ))}
-            </RadioGroup>
-          </FormItem>
-          <FormItem label={t('common:videoDisplay')} {...FORM_ITEM_LAYOUT}>
-            <Switch
-              size="small"
-              checked={showVideo}
-              disabled={shouldDisableVideo(tracks[0].sourceUrl)}
-              onChange={handleShowVideoChanged}
-              />
-          </FormItem>
-          <FormItem label={t('common:posterImageUrl')} {...FORM_ITEM_LAYOUT}>
-            <UrlInput
-              value={posterImage.sourceUrl}
-              allowedSourceTypes={ensureIsExcluded(Object.values(SOURCE_TYPE), SOURCE_TYPE.youtube)}
-              disabled={shouldDisableVideo(tracks[0].sourceUrl) || !showVideo}
-              onChange={handlePosterImageSourceUrlChange}
-              />
-          </FormItem>
-          <FormItem
-            label={<Info tooltip={t('common:widthInfo')}>{t('common:width')}</Info>}
-            {...FORM_ITEM_LAYOUT}
-            >
-            <ObjectWidthSlider value={width} onChange={handleWidthChanged} />
-          </FormItem>
-          <FormItem
-            label={<Info tooltip={t('common:multitrackInitialVolumeInfo')}>{t('common:initialVolume')}</Info>}
-            {...FORM_ITEM_LAYOUT}
-            >
-            <MediaVolumeSlider
-              value={initialVolume}
-              useValueLabel
-              useButton={false}
-              onChange={handleInitialVolumeChange}
-              />
-          </FormItem>
+          <PlayerSettingsEditor
+            content={content}
+            disableVideo={disableVideo}
+            onContentChange={handlePlayerSettingsContentChange}
+            />
         </ItemPanel>
 
         <ItemPanel header={t('common:trackMixer')}>
@@ -242,7 +168,7 @@ function MultitrackMediaEditor({ content, onContentChanged }) {
               posterImageUrl={getAccessibleUrl({ url: posterImage.sourceUrl, cdnRootUrl: clientConfig.cdnRootUrl })}
               screenWidth={50}
               selectedVolumePresetIndex={selectedVolumePresetIndex}
-              showVideo={showVideo}
+              showVideo={!disableVideo && showVideo}
               showTrackMixer={false}
               sources={sources}
               volumePresets={volumePresets}
